feat(data-table): implement tree expansion logic and refactor selection handling
